refactor: Move view and command registration out of main extension file
