+++ src/commands/index.ts
@@ -1,0 +1,3 @@
+export * from './init'
+export * from './usagesNodeClicked'
+export * from './openLink'
