+++ src/commands/init/constants.ts
@@ -1,0 +1,1 @@
+export const COMMAND_INIT = 'devcycle-feature-flags.init'
