+++ src/commands/logout/constants.ts
@@ -1,0 +1,1 @@
+export const COMMAND_LOGOUT = 'devcycle-feature-flags.logout'
