+++ src/commands/logout/index.ts
@@ -1,0 +1,2 @@
+export * from './registerLogoutCommand'
+export * from './constants'
