+++ src/commands/showReference/index.ts
@@ -1,0 +1,2 @@
+export * from './registerShowReferenceCommand'
+export * from './constants'
