+++ src/views/login/registerLoginViewProvider.ts
@@ -1,0 +1,14 @@
+import * as vscode from 'vscode'
+import { LoginViewProvider } from './LoginViewProvider'
+
+export async function registerLoginViewProvider(context: vscode.ExtensionContext) {
+  const loginViewProvider = new LoginViewProvider(context.extensionUri)
+  context.subscriptions.push(
+    vscode.window.registerWebviewViewProvider(
+      'devcycle-sidebar',
+      loginViewProvider
+    ),
+  )
+
+  return loginViewProvider
+}
