+++ src/views/startup/index.ts
@@ -1,0 +1,2 @@
+export * from './StartupViewProvider'
+export * from './registerStartupViewProvider'
