+++ src/views/usages/index.ts
@@ -1,0 +1,2 @@
+export * from './registerUsagesViewProvider'
+export * from './UsagesTree'
